+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 ### Fixes
  * Save As now selects correct file extension by default.
+ * Saving an untitled file now correctly opens the Save As dialog instead of just saving the file to `untitled n`.
 
 ## Version 0.4.0-alpha
 ### Features